clean up n refactor
